Move some missed functions in search split
